Implement Trigger system to clean up event handler
